# frontend/src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Food Recommender API",
...
        "contact": {
            "name": "Food Recommender Team",
            "email": "support@foodrecommender.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": "https://api.foodrecommender.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/user_profile/{user_id}": {
            "get": {
                "tags": [
                    "user_profile"
                ],
                "summary": "Get User Taste Profile",
                "description": "Retrieve a user's complete taste profile including preferences and food history",
                "operationId": "get_user_profile_user_profile__user_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Taste profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TasteProfileResponse"
                                },
                                "example": {
                                    "profile": {
                                        "user_id": "uuid-string",
                                        "dietary_restrictions": [
                                            "pescetarian",
                                            "gluten-free"
                                        ],
                                        "cuisine_preferences": [
                                            "italian",
                                            "mexican",
                                            "thai"
                                        ],
                                        "flavor_profile": "spicy",
                                        "liked_foods": [
                                            "pasta",
                                            "tacos",
                                            "pad thai"
                                        ],
                                        "disliked_foods": [
                                            "mushrooms",
                                            "olives"
                                        ],
                                        "favorite_restaurants": [
                                            "restaurant_id_1",
                                            "restaurant_id_2"
                                        ],
                                        "price_range_preference": "medium",
                                        "meal_time_preferences": {
                                            "breakfast": [
                                                "light"
                                            ],
                                            "dinner": [
                                                "hearty"
                                            ]
                                        },
                                        "updated_at": "2024-01-01T00:00:00"
                                    },
                                    "recommendations_count": 15,
                                    "last_activity": "2024-01-01T00:00:00"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Not authorized to access this profile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user_profile"
                ],
                "summary": "Update User Taste Profile",
                "description": "Update a user's taste profile with new preferences and food history",
                "operationId": "update_user_profile_user_profile__user_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "User Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTasteProfile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TasteProfileResponse"
                                },
                                "example": {
                                    "profile": {
                                        "user_id": "uuid-string",
                                        "dietary_restrictions": [
                                            "vegetarian"
                                        ],
                                        "cuisine_preferences": [
                                            "italian",
                                            "mexican"
                                        ],
                                        "flavor_profile": "mild",
                                        "liked_foods": [
                                            "pasta",
                                            "pizza"
                                        ],
                                        "disliked_foods": [
                                            "mushrooms"
                                        ],
                                        "favorite_restaurants": [
                                            "restaurant_id_1"
                                        ],
                                        "price_range_preference": "medium",
                                        "meal_time_preferences": {
                                            "lunch": [
                                                "quick"
                                            ]
                                        },
                                        "updated_at": "2024-01-01T12:00:00"
                                    },
                                    "recommendations_count": 8,
                                    "last_activity": "2024-01-01T12:00:00"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Not authorized to update this profile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user_profile"
                ],
                "summary": "Delete User Taste Profile",
                "description": "Permanently delete a user's taste profile and all associated data",
                "operationId": "delete_user_profile_user_profile__user_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Profile deleted successfully"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Not authorized to delete this profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Profile not found"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CuisinePreference": {
                "properties": {
                    "cuisine_type": {
                        "type": "string",
                        "title": "Cuisine Type"
                    },
                    "preference_level": {
                        "type": "integer",
                        "title": "Preference Level"
                    }
                },
                "type": "object",
                "required": [
                    "cuisine_type",
                    "preference_level"
                ],
                "title": "CuisinePreference"
            },
            "DietaryRestriction": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "severity": {
                        "type": "string",
                        "title": "Severity"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "severity"
                ],
                "title": "DietaryRestriction"
            },
            "FlavorProfile": {
                "properties": {
                    "spicy_tolerance": {
                        "type": "integer",
                        "title": "Spicy Tolerance"
                    },
                    "sweet_preference": {
                        "type": "integer",
                        "title": "Sweet Preference"
                    },
                    "salty_preference": {
                        "type": "integer",
                        "title": "Salty Preference"
                    },
                    "sour_preference": {
                        "type": "integer",
                        "title": "Sour Preference"
                    },
                    "umami_preference": {
                        "type": "integer",
                        "title": "Umami Preference"
                    },
                    "bitter_tolerance": {
                        "type": "integer",
                        "title": "Bitter Tolerance"
                    }
                },
                "type": "object",
                "required": [
                    "spicy_tolerance",
                    "sweet_preference",
                    "salty_preference",
                    "sour_preference",
                    "umami_preference",
                    "bitter_tolerance"
                ],
                "title": "FlavorProfile"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TasteProfileResponse": {
                "properties": {
                    "profile": {
                        "$ref": "#/components/schemas/UserTasteProfile"
                    },
                    "recommendations_count": {
                        "type": "integer",
                        "title": "Recommendations Count"
                    },
                    "last_activity": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Activity"
                    }
                },
                "type": "object",
                "required": [
                    "profile",
                    "recommendations_count"
                ],
                "title": "TasteProfileResponse"
            },
...
            "UpdateTasteProfile": {
                "properties": {
                    "dietary_restrictions": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/DietaryRestriction"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dietary Restrictions"
                    },
                    "cuisine_preferences": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/CuisinePreference"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cuisine Preferences"
                    },
                    "flavor_profile": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/FlavorProfile"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "liked_foods": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/FoodItem"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Liked Foods"
                    },
                    "disliked_foods": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/FoodItem"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Disliked Foods"
                    },
                    "favorite_restaurants": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Favorite Restaurants"
                    },
                    "price_range_preference": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Range Preference"
                    },
                    "meal_time_preferences": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Meal Time Preferences"
                    }
                },
                "type": "object",
                "title": "UpdateTasteProfile"
            },
...
            "UserTasteProfile": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "dietary_restrictions": {
                        "items": {
                            "$ref": "#/components/schemas/DietaryRestriction"
                        },
                        "type": "array",
                        "title": "Dietary Restrictions",
                        "default": []
                    },
                    "cuisine_preferences": {
                        "items": {
                            "$ref": "#/components/schemas/CuisinePreference"
                        },
                        "type": "array",
                        "title": "Cuisine Preferences",
                        "default": []
                    },
                    "flavor_profile": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/FlavorProfile"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "liked_foods": {
                        "items": {
                            "$ref": "#/components/schemas/FoodItem"
                        },
                        "type": "array",
                        "title": "Liked Foods",
                        "default": []
                    },
                    "disliked_foods": {
                        "items": {
                            "$ref": "#/components/schemas/FoodItem"
                        },
                        "type": "array",
                        "title": "Disliked Foods",
                        "default": []
                    },
                    "favorite_restaurants": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Favorite Restaurants",
                        "default": []
                    },
                    "price_range_preference": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Range Preference"
                    },
                    "meal_time_preferences": {
                        "additionalProperties": {
                            "items": {
                                "type": "string"
                            },
                            "type": "array"
                        },
                        "type": "object",
                        "title": "Meal Time Preferences",
                        "default": {}
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "updated_at"
                ],
                "title": "UserTasteProfile"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}